Separate tests for each config
Besides aligning to the convention that test files match the name of the file they tests,
it generates separate snapshots for each of the configuration, which will help looking at the diff on GitHub.
